Extract admin setup bootstrap to feature module
